requirements.txt
asgiref==3.8.1
Django==5.1.2
django-allauth==65.0.2
django-jazzmin==3.0.1
django-cleanup==9.0.0
django-htmx==1.19.0
dj-database-url==2.2.0
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
tzdata==2024.2
python-dotenv==0.21.1
requests==2.31.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0